Factor out extractPackage
