+++ api/src/lib/expand-function/expand-function.ts
@@ -8,7 +8,11 @@ import {
   getDefinitionText,
   getParentImportDeclaration,
 } from "./ast-helpers/index.js";
-import { contextForImport, resolveModulePath } from "./imports/index.js";
+import {
+  contextForImport,
+  extractPackageName,
+  resolveModulePath,
+} from "./imports/index.js";
 import { searchFunction } from "./search-function/index.js";
 import {
   getFileUri,
@@ -383,47 +387,4 @@ async function extractContext(
   }
 
   return context;
-}
-
-function extractPackageName(uri: string): string | null {
-  try {
-    const url = new URL(uri);
-    const path = decodeURIComponent(url.pathname); // Decode URI components
-
-    const nodeModulesIndex = path.indexOf("node_modules");
-
-    if (nodeModulesIndex === -1) {
-      return null;
-    }
-
-    const packagePath = path.substring(
-      nodeModulesIndex + "node_modules/".length,
-    );
-
-    // Handle pnpm structure
-    const pnpmIndex = packagePath.indexOf(".pnpm/");
-    if (pnpmIndex !== -1) {
-      const pnpmPath = packagePath.substring(pnpmIndex + ".pnpm/".length);
-      const parts = pnpmPath.split("/");
-
-      // Handle scoped packages
-      if (parts[0].startsWith("@") && parts.length > 1) {
-        return `${parts[0]}/${parts[1]}`;
-      }
-
-      return parts[0];
-    }
-
-    const parts = packagePath.split("/");
-
-    // Handle scoped packages
-    if (parts[0].startsWith("@") && parts.length > 1) {
-      return `${parts[0]}/${parts[1]}`;
-    }
-
-    return parts[0];
-  } catch (error) {
-    logger.error("[extractPackageName] Error parsing URI:", uri, error);
-    return null;
-  }
 }
