+++ api/src/lib/expand-function/imports/index.ts
@@ -1,2 +1,3 @@
 export * from "./context-for-import.js";
 export { resolveModulePath } from "./resolve-imports.js";
+export { extractPackageName } from "./extract-package-name.js";
